# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "URL Shortener API with analytics and link management",
        "title": "Pendek.in API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@pendek.in"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/analytics/": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detailed analytics including device breakdowns, countries, traffic sources, and browser usage",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analytics"
                ],
                "summary": "Get analytics data",
                "parameters": [
                    {
                        "enum": [
                            "7d",
                            "30d",
                            "90d",
                            "all"
                        ],
                        "type": "string",
                        "default": "30d",
                        "description": "Time range",
                        "name": "range",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/responses.AnalyticsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/analytics/dashboard": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get dashboard overview including total clicks, active links, top link, and recent links",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analytics"
                ],
                "summary": "Get dashboard data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/responses.DashboardResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "responses.AnalyticOverview": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "value": {
                    "type": "integer"
                }
            }
        },
        "responses.AnalyticsResponse": {
            "type": "object",
            "properties": {
                "avg_daily_click": {
                    "type": "integer"
                },
                "browser_usages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.TypeValue"
                    }
                },
                "device_breakdowns": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.TypeValue"
                    }
                },
                "from_date": {
                    "type": "string"
                },
                "overviews": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.AnalyticOverview"
                    }
                },
                "time_range": {
                    "type": "string"
                },
                "to_date": {
                    "type": "string"
                },
                "top_countries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.TypeValue"
                    }
                },
                "top_link": {
                    "$ref": "#/definitions/responses.TopLink"
                },
                "total_active_links": {
                    "type": "integer"
                },
                "total_clicks": {
                    "type": "integer"
                },
                "traffic_sources": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.TypeValue"
                    }
                }
            }
        },
        "responses.BaseResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "responses.DashboardResponse": {
            "type": "object",
            "properties": {
                "overviews": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.AnalyticOverview"
                    }
                },
                "recents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.LinkResponse"
                    }
                },
                "top_link": {
                    "$ref": "#/definitions/responses.TopLink"
                },
                "total_active_links": {
                    "type": "integer"
                },
                "total_clicks": {
                    "type": "integer"
                }
            }
        },
        "responses.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "responses.LinkResponse": {
            "type": "object",
            "properties": {
                "click_count": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "custom_short_code": {
                    "type": "string"
                },
                "expired_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "original_url": {
                    "type": "string"
                },
                "short_code": {
                    "type": "string"
                }
            }
        },
...
        "responses.TopLink": {
            "type": "object",
            "properties": {
                "link": {
                    "$ref": "#/definitions/responses.LinkResponse"
                },
                "total_clicks": {
                    "type": "integer"
                }
            }
        },
        "responses.TypeValue": {
            "type": "object",
            "properties": {
                "type": {
                    "type": "string"
                },
                "value": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}